+++ src/test/mod.rs
@@ -1,5 +1,4 @@
 use rstest::rstest;
-use serde_json::json;
 use speculoos::assert_that;
 use std::path::Path;
 
@@ -8,6 +7,7 @@ use crate::cli::Cli;
 #[rstest]
 #[case("dotfiles01", "toml")]
 #[case("dotfiles02", "json")]
+#[case("dotfiles03", "force")]
 fn read_config_formats(#[case] dotfiles_path: &str, #[case] expexted: &str) {
   let mut cli = Cli {
     dry_run: true,
@@ -18,12 +18,12 @@ fn read_config_formats(#[case] dotfiles_path: &str, #[case] expexted: &str) {
 
   let config = super::read_config(&cli).unwrap();
 
-  assert_that!(config.variables["test01"]).is_equal_to(json!("yaml"));
+  assert_that!(config.variables["test01"]).is_equal_to(&"yaml".into());
 
   cli.dotfiles = Some(Path::new(file!()).parent().unwrap().join("data").join(dotfiles_path).into());
 
   let config = super::read_config(&cli).unwrap();
 
-  assert_that!(config.variables["test01"]).is_equal_to(json!("yaml"));
-  assert_that!(config.variables["test02"]).is_equal_to(json!(expexted));
+  assert_that!(config.variables["test01"]).is_equal_to(&"yaml".into());
+  assert_that!(config.variables["test02"]).is_equal_to(&expexted.into());
 }